Change Tx History to be a seperate page
