+++ packages/widget-v2/src/pages/SwapPage/SwapPage.tsx
@@ -29,7 +29,6 @@ import { useModal } from "@/components/Modal";
 import { WalletSelectorModal } from "@/modals/WalletSelectorModal/WalletSelectorModal";
 import { currentPageAtom, Routes } from "@/state/router";
 import { useInsufficientSourceBalance } from "./useSetMaxAmount";
-import { TransactionHistoryModal } from "@/modals/TransactionHistoryModal/TransactionHistoryModal";
 import { errorAtom, ErrorType } from "@/state/errorPage";
 import { ConnectedWalletContent } from "./ConnectedWalletContent";
 import { useSourceAccount } from "@/hooks/useSourceAccount";
@@ -54,7 +53,6 @@ export const SwapPage = () => {
   const tokenAndChainSelectorModal = useModal(TokenAndChainSelectorModal);
   const selectWalletmodal = useModal(WalletSelectorModal);
   const setCurrentPage = useSetAtom(currentPageAtom);
-  const historyModal = useModal(TransactionHistoryModal);
   const insufficientBalance = useInsufficientSourceBalance();
   const setSwapExecutionState = useSetAtom(setSwapExecutionStateAtom);
   const setError = useSetAtom(errorAtom);
@@ -239,7 +237,7 @@ export const SwapPage = () => {
           leftButton={{
             label: "History",
             icon: ICONS.history,
-            onClick: () => historyModal.show(),
+            onClick: () => setCurrentPage(Routes.TransactionHistoryPage),
           }}
           rightContent={<ConnectedWalletContent />}
         />
