+++ packages/widget-v2/src/state/router.ts
@@ -3,6 +3,7 @@ import { atom } from "jotai";
 export enum Routes {
   SwapPage,
   SwapExecutionPage,
+  TransactionHistoryPage,
 }
 
 export const currentPageAtom = atom<Routes>(Routes.SwapPage);
