+++ packages/widget-v2/src/stories/tx_history.json
